Move save and cleanup to utils
