+++ prepare_dataset.py
@@ -91,29 +91,6 @@ def process_image(image, filename):
                 tile_index += 1
 
 
-def cleanup(save_list):
-    for img in save_list:
-        img[0].close()
-
-
-def save():
-    save_start = int(time.time())
-    print("Saving pictures (all at once, might take a while)...")
-    print("Saving LR...")
-    for img in lr_save_list:
-        img[0].save(img[1], "PNG", icc_profile='')
-    print("Freeing resources...")
-    cleanup(lr_save_list)
-    print("Saving HR...")
-    for img in hr_save_list:
-        img[0].save(img[1], "PNG", icc_profile='')
-    print("Freeing resources...")
-    cleanup(hr_save_list)
-
-    save_end = int(time.time())
-    print("Time taken: {} - {} = {}".format(save_start, save_end, save_start - save_end))
-
-
 def main():
     print("Splitting dataset pictures...")
     rgb_index = 0
@@ -132,7 +109,7 @@ def main():
                 picture.close()
                 index += 1
     if use_ram:
-        save()
+        extrasUtil.save(lr_save_list, hr_save_list)
     print("{} pictures were converted to RGB.".format(rgb_index))
